misc/HtmlTestSuite/results/test-0228.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf8"/>
<title>CSS Content Visibility: offscreen c-v auto content is relevant when in top layer</title>
<link href="mailto:mrobinson@igalia.com" rel="author" title="Martin Robinson"/>
<link href="https://drafts.csswg.org/css-contain/#content-visibility" rel="help"/>
<meta content="offscreen c-v auto content is relevant when in top layer" name="assert"/>

<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>

<style>
#inner {
    content-visibility: auto;
    contain-intrinsic-size: 100px 100px;
}
</style>

</head><body><dialog id="dialog">
    <div id="spacer" style="height: 100000px;"/>
    <div id="inner">
        <div style="height: 200px; width: 200px;">content</div>
    </div>
</dialog>

<script>
test(() =&gt; {
    dialog.show();

    assert_equals(inner.getBoundingClientRect().height, 100);

    dialog.close();
    dialog.showModal();

    // Even though the element with `content-visibility: auto` is offscreen
    // it should be considered relevant content because it is now in the top
    // layer.
    assert_equals(inner.getBoundingClientRect().height, 200);
});
</script>
</body></html>
